+++ src/compas_fab/robots/__init__.py
@@ -79,6 +79,5 @@ from .units import *
 from .robot import *
 from .semantics import *
 from .trajectory import *
-from .ros_fileserver_loader import *
 
 __all__ = [name for name in dir() if not name.startswith('_')]